+++ example/integration_test/_test_imports.dart
@@ -1,0 +1,7 @@
+//
+export '../test/src/_test_imports.dart';
+
+// Individual Tests
+export 'src/animals_test.dart';
+export 'src/counter_test.dart';
+export 'src/words_test.dart';