+++ example/lib/src/app/view/menu/app_menu.dart
@@ -7,14 +7,13 @@ import '/src/view.dart';
 class AppMenu extends AppPopupMenu {
   /// Only one instance of the class
   factory AppMenu() => _this ??= AppMenu._();
-  static AppMenu? _this;
-
   AppMenu._()
       : _con = ExampleAppController(),
         super(
           key: const Key('appMenuButton'),
           controller: AppMenuController(),
         );
+  static AppMenu? _this;
 
   /// The App's controller
   final ExampleAppController _con;
@@ -66,7 +65,7 @@ class AppMenu extends AppPopupMenu {
 
   /// When an menu item is selected
   @override
-  void onSelected(String value) async {
+  Future<void> onSelected(String value) async {
     switch (value) {
       case 'interface':
         _con.changeUI();
@@ -75,10 +74,10 @@ class AppMenu extends AppPopupMenu {
         await _con.changeApp();
         break;
       case 'locale':
-        _con.changeLocale();
+        await _con.changeLocale();
         break;
       case 'color':
-        _con.changeColor();
+        await _con.changeColor();
         break;
       case 'about':
         _con.aboutApp();
@@ -90,7 +89,8 @@ class AppMenu extends AppPopupMenu {
 
 /// Uses the String type as menu options
 class AppMenuController extends AppPopupMenuController {
+  ///
   factory AppMenuController() => _this ??= AppMenuController._();
-  static AppMenuController? _this;
   AppMenuController._();
+  static AppMenuController? _this;
 }