+++ example/lib/src/app/view/menu/radio_material.dart
@@ -19,7 +19,7 @@ class MaterialVersionRadioButtons extends RadioButtons<bool> {
       MaterialController(),
       mainAxisSize: MainAxisSize.min,
     );
-    List<Widget> widgets = [Text('Material Ver.'.tr)];
+    final List<Widget> widgets = [Text('Material Ver.'.tr)];
     widgets.addAll(radios);
     return Row(
       mainAxisAlignment: MainAxisAlignment.spaceBetween,
@@ -54,6 +54,7 @@ class MaterialVersionRadioButtons extends RadioButtons<bool> {
 
 /// Supplies the initial value and callbacks when a new selection
 class MaterialController extends RadioButtonsController<bool> {
+  ///
   factory MaterialController() => _this ??= MaterialController._();
   MaterialController._();
   static MaterialController? _this;
@@ -101,10 +102,13 @@ class MaterialController extends RadioButtonsController<bool> {
     // The radio buttons group value
     groupValue = initValue;
   }
-
+  ///
   bool initValue = false;
 
+  ///
   static ThemeData? theme03Data; // Material3
+
+  ///
   static ThemeData? theme02Data; // Material2
 
   @override