+++ example/test/src/_test_imports.dart
@@ -17,9 +17,5 @@ export 'tests/menu/open_drawer.dart';
 export 'tests/menu/open_menu.dart';
 
 /// Navigation Tests
-export '../test_navigator_mixin.dart';
+export 'test_navigator_mixin.dart';
 
-/// Individual Tests
-export 'tests/counter_test.dart';
-export 'tests/words_test.dart';
-// export 'tests/contacts_test.dart';